migrado pagina de usuarios

+++ backend/core/helpers.go
@@ -201,8 +201,8 @@ func SUnixTimeMilli() int64 {
 func UnixToSunix(unixTime int64) int32 {
 	return int32((unixTime - 1e9) / 2)
 }
-func SunixToUnix(sunixTime int32) int32 {
-	return int32((sunixTime + 1e9) * 2)
+func SunixToUnix(sunixTime int32) int64 {
+	return (int64(sunixTime) + 1e9) * 2
 }
 
 // Makes an UUID based on SUnixTime in milliseconds and random 3 last digits
